requirements.txt
pandas 
numpy
matplotlib
japanize_matplotlib
seaborn
scikit-learn
streamlit
requests
python-dotenv
